requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
pypdf==5.0.1
openai==1.51.2

slowapi==0.1.9
aiofiles==24.1.0